refactor(BA-6483): move AppConfigPolicyAdminProcessors to admin_processors.py
Mirror the admin_service.py file split for the processor package.
