+++ src/ai/backend/manager/services/factory.py
@@ -6,11 +6,13 @@ from ai.backend.manager.repositories.resource_allocation.repository import (
 )
 from ai.backend.manager.services.agent.processors import AgentProcessors
 from ai.backend.manager.services.agent.service import AgentService
+from ai.backend.manager.services.app_config_policy.admin_processors import (
+    AppConfigPolicyAdminProcessors,
+)
 from ai.backend.manager.services.app_config_policy.admin_service import (
     AppConfigPolicyAdminService,
 )
 from ai.backend.manager.services.app_config_policy.processors import (
-    AppConfigPolicyAdminProcessors,
     AppConfigPolicyProcessors,
 )
 from ai.backend.manager.services.app_config_policy.service import AppConfigPolicyService
